Moved fullscreen logic to screen_handling in prep for a fix.
It needs to be further moved to WindowManager instead of screen_handling, to be beside `bool setWindowSize`.
