+++ src/kre/WindowManager.hpp
@@ -34,12 +34,6 @@
 
 namespace KRE
 {
-	enum class FullScreenMode {
-		WINDOWED,
-		FULLSCREEN_WINDOWED,
-		FULLSCREEN_EXCLUSIVE,
-	};
-
 	struct WindowMode
 	{
 		int width;
